backend/requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.1
factory-boy>=3.3
faker>=22.0
django-debug-toolbar>=4.3
ruff>=0.2
ipython>=8.20
